+++ CleverPush/CleverPush/Source/UNUserNotificationCenter+CleverPush.m
@@ -98,11 +98,11 @@ didReceiveNotificationResponse:(UNNotificationResponse *)response
 withCompletionHandler:(void(^)())completionHandler {
     NSLog(@"CleverPush cleverPushUserNotificationCenter didReceiveNotificationResponse");
     
+    /*
     if (![CleverPush channelId]) {
-
-//        return;
-
+        return;
     }
+    */
     
     if ([CleverPushUNUserNotificationCenter isDismissEvent:response]) {
         [CleverPush updateBadge:nil];